Added Magma, and altered Semigroup so it's based on Magma.
